refactor: 박스스코어 쿼리 코드 상수화(#169)
