+++ src/features/game/components/table/BattingRecordTable.tsx
@@ -1,6 +1,6 @@
 import { battingRecordColumns } from '@/constants/columns/batting-record-columns';
 import DataTable from '@/features/common/DataTable';
-import type { Batter } from '../../types/BoxscoreData';
+import type { Batter } from '../../types/boxscore';
 
 interface BattingRecordTableProps {
   data: Batter[] | undefined;
